feat(controls-island): implement auto-hide and auto-collapse composables for improved UI behavior
