+++ packages/stage-ui/src/composables/index.ts
@@ -13,6 +13,7 @@ export * from './use-inference-status'
 export * from './use-lamp-flicker-animation'
 export * from './use-linked-accounts'
 export * from './use-model-preload'
+export * from './use-controls-island-autohide'
 export * from './use-optimistic'
 export * from './use-scroll-to-hash'
 export * from './vision'
